# docs/public/specs/openapi.yaml
openapi: 3.1.0
info:
  title: Macro Deck plugin protocol - REST surface
  version: "1"
  description: >-
    Registration, session exchange and pre-registration discovery for out-of-process Macro Deck
    plugins. The wire envelope carried over the WebSocket upgrade at the end of this surface is
    specified in asyncapi.yaml, not here. Paths are deliberately unversioned - the protocol version
    is negotiated over POST /sessions, never encoded in the URL.

# Drift anchor: MacroDeck.Plugin.Protocol.Versioning.ProtocolVersions
...
servers:
  - url: /
security: []
paths:
...
  /api/plugins/pairing/{requestId}:
    get:
      operationId: getPluginPairingStatus
      summary: Poll a pairing request's status
      description: >-
        An unknown or already-pruned request id deliberately answers 200 with status "expired" rather
        than 404, so the client state machine stays total and this endpoint is not an existence oracle.
      security: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The current status of the pairing request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginPairingStatusResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/plugins/pairing/{requestId}/redemption:
    post:
      operationId: redeemPluginPairingRequest
      summary: Redeem an approved pairing request for a plugin credential
      description: >-
        Unknown, expired, unapproved, already-redeemed and wrong-verifier request/verifier combinations
        all answer 401 indistinguishably, deliberately - the same oracle argument as registration
        authentication.
      security: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PluginPairingRedemptionRequest"
      responses:
        "201":
          description: The pairing request was redeemed. The secret is shown exactly once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginRegistrationResponse"
        "400":
          $ref: "#/components/responses/InvalidPayload"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/plugins/sessions:
    post:
      operationId: createPluginSession
      summary: Exchange a plugin secret for a session token
      description: >-
        Authenticated by the plugin id and secret headers. Negotiation happens here, once, and is
        authoritative - the subsequent session.hello over the WebSocket only asserts the outcome,
        it never re-negotiates.
      security:
        - pluginCredential: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PluginSessionRequest"
      responses:
        "201":
          description: The session was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginSessionResponse"
        "400":
          $ref: "#/components/responses/InvalidPayload"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ProtocolVersionUnsupported"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/plugins/sessions/{sessionId}:
    delete:
      operationId: closePluginSession
      summary: Voluntary session teardown
      description: Makes the session non-resumable at once.
      security:
        - sessionBearer: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: The session was closed.
        "400":
          $ref: "#/components/responses/InvalidPayload"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    enrollmentToken:
      type: apiKey
      in: header
      name: X-MacroDeck-Enrollment-Token
    pluginCredential:
      type: apiKey
      in: header
      name: X-MacroDeck-Plugin-Secret
      description: >-
        Sent alongside the X-MacroDeck-Plugin-Id header. Session exchange only - never sent on the
        WebSocket upgrade.
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Authorization: Bearer <sessionToken>, scope \"plugin\"."
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Authorization: Bearer <accessToken>, scope \"admin\"."
  responses:
    InvalidPayload:
      description: The request body does not match the expected shape.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProtocolError"
    Unauthenticated:
      description: >-
        Authentication failed. An unknown id and a wrong secret are indistinguishable in the
        response, the same oracle argument as device login.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProtocolError"
    Forbidden:
      description: >-
        The caller did not arrive on a loopback remote address. Plugin endpoints are served on both
        the public and private listeners, but only local processes may reach them. On the pairing
        endpoints, on POST /api/plugins/registration, and on POST /api/plugins/sessions for a plugin
        holding a development credential, this status also means Developer Mode is switched off in
        the desktop app - such a refusal carries "reason": "developer_mode_disabled" in the error's
        details, and GET /api/plugins/protocol reports the switch ahead of time.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProtocolError"
...
    ProtocolVersionUnsupported:
      description: The requested protocol version is not supported.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProtocolError"
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After:
          required: true
          schema:
            type: integer
            format: int32
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProtocolError"
  schemas:
    ProtocolErrorCode:
      type: string
      description: One of the codes v1 speaks. Drift anchor - see x-macrodeck-error-codes above.
      enum:
        - PROTOCOL_VERSION_UNSUPPORTED
        - UNKNOWN_MESSAGE_TYPE
        - MALFORMED_ENVELOPE
        - INVALID_PAYLOAD
        - UNAUTHENTICATED
        - PLUGIN_ALREADY_REGISTERED
        - SESSION_EXPIRED
        - SESSION_NOT_RESUMABLE
        - SESSION_REPLACED
        - SESSION_NOT_FOUND
        - CAPABILITY_UNSUPPORTED
        - CAPABILITY_UNAVAILABLE
        - PAYLOAD_TOO_LARGE
        - ASSET_TOO_LARGE
        - QUEUE_OVERFLOW
        - RATE_LIMITED
        - TIMEOUT
        - CANCELLED
        - CORRELATION_UNKNOWN
        - DUPLICATE_IDEMPOTENCY_KEY
        - INTERNAL_ERROR
        - ADB_NOT_ENABLED
        - ADB_NOT_ALLOWED
        - ADB_FAILED
        - UI_RESOURCE_QUOTA_EXCEEDED
        - PLUGIN_ICON_NOT_FOUND
        - ICON_PACK_INVALID
        - ICON_PACK_SYNC_NOT_ALLOWED
    ProtocolError:
      type: object
      description: Mirrors MacroDeck.Plugin.Protocol.Errors.ProtocolError.
      required: [code, message, retryable]
      properties:
        code:
          $ref: "#/components/schemas/ProtocolErrorCode"
        message:
          type: string
          description: >-
            Default English text keyed by code (MacroDeck.Plugin.Protocol.Errors.ProtocolErrorMessages).
            The UI localises from the code, not this string.
        details:
          type: object
          additionalProperties:
            type: string
        retryable:
          type: boolean
    PluginProtocolLimitsDescriptor:
      type: object
      description: Mirrors MacroDeck.Plugin.Protocol.Handshake.PluginProtocolLimitsDescriptor.
      required:
        - maxMessageBytes
        - maxAssetBytes
        - maxAssetChunkBytes
        - maxInboundQueueDepth
        - maxOutboundQueueDepth
        - queueHighWatermark
        - queueLowWatermark
        - maxConcurrentInvocations
        - maxDeclaredCapabilities
        - maxIdempotencyKeyLength
        - maxErrorMessageLength
        - maxErrorDetailEntries
        - maxJsonDepth
        - maxSessionsPerPlugin
      properties:
        maxMessageBytes:
          type: integer
        maxAssetBytes:
          type: integer
        maxAssetChunkBytes:
          type: integer
        maxInboundQueueDepth:
          type: integer
        maxOutboundQueueDepth:
          type: integer
        queueHighWatermark:
          type: integer
        queueLowWatermark:
          type: integer
        maxConcurrentInvocations:
          type: integer
        maxDeclaredCapabilities:
          type: integer
        maxIdempotencyKeyLength:
          type: integer
        maxErrorMessageLength:
          type: integer
        maxErrorDetailEntries:
          type: integer
        maxJsonDepth:
          type: integer
        maxSessionsPerPlugin:
          type: integer
    PluginProtocolTimeoutsDescriptor:
      type: object
      description: >-
        Mirrors MacroDeck.Plugin.Protocol.Handshake.PluginProtocolTimeoutsDescriptor. Each field is
        a .NET TimeSpan serialized in its default "c" format (e.g. "00:00:10" for ten seconds).
      required:
        - handshake
        - defaultRequest
        - capabilityInvoke
        - assetUpload
        - keepAliveInterval
        - keepAliveTimeout
        - sessionResumeWindow
        - gracefulClose
      properties:
        handshake:
          type: string
        defaultRequest:
          type: string
        capabilityInvoke:
          type: string
        assetUpload:
          type: string
        keepAliveInterval:
          type: string
        keepAliveTimeout:
          type: string
        sessionResumeWindow:
          type: string
        gracefulClose:
          type: string
...
    PluginPairingStatusResponse:
      type: object
      description: Mirrors MacroDeck.Plugin.Protocol.Handshake.PluginPairingStatusResponse.
      required: [status, expiresAt]
      properties:
        status:
          type: string
          description: Mirrors MacroDeck.Plugin.Protocol.Handshake.PluginPairingStatuses.All.
          enum: [pending, approved, rejected, expired]
        expiresAt:
          type: string
          format: date-time
    PluginPairingRedemptionRequest:
      type: object
      description: Mirrors MacroDeck.Plugin.Protocol.Handshake.PluginPairingRedemptionRequest.
      required: [codeVerifier]
      properties:
        codeVerifier:
          type: string
...
    PluginRegistrationResponse:
      type: object
      description: Mirrors MacroDeck.Plugin.Protocol.Handshake.PluginRegistrationResponse.
      required: [pluginId, pluginSecret]
      properties:
        pluginId:
          type: string
        pluginSecret:
          type: string
          description: Shown exactly once. The host stores only its hash.
    ProtocolVersionRange:
      type: object
      description: Mirrors MacroDeck.Plugin.Protocol.Versioning.ProtocolVersionRange.
      required: [minimum, maximum]
      properties:
        minimum:
          type: integer
        maximum:
          type: integer
    CapabilityVersionRange:
      type: object
      description: Mirrors MacroDeck.Plugin.Protocol.Versioning.CapabilityVersionRange.
      required: [minimum, maximum]
      properties:
        minimum:
          type: integer
        maximum:
          type: integer
    DeclaredCapability:
      type: object
      description: Mirrors MacroDeck.Plugin.Protocol.Handshake.DeclaredCapability.
      required: [kind, localId, versionRange]
      properties:
        kind:
          type: string
          description: One of x-macrodeck-capability-kinds.
        localId:
          type: string
        versionRange:
          $ref: "#/components/schemas/CapabilityVersionRange"
        displayName:
          type: string
    CapabilityNegotiationResult:
      type: object
      description: Mirrors MacroDeck.Plugin.Protocol.Versioning.CapabilityNegotiationResult.
      required: [kind, accepted]
      properties:
        kind:
          type: string
        accepted:
          type: boolean
        negotiatedVersion:
          type: integer
        rejectionReason:
          type: string
    PluginSessionRequest:
      type: object
      description: Mirrors MacroDeck.Plugin.Protocol.Handshake.PluginSessionRequest.
      required: [requestedVersion, capabilities]
      properties:
        requestedVersion:
          $ref: "#/components/schemas/ProtocolVersionRange"
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/DeclaredCapability"
        declaredName:
          type: string
          nullable: true
          maxLength: 128
          description: >-
            The plugin's own display name for this session, as declared by the plugin process itself. Optional and
            additive - names the current session only, never the plugin's enrollment record.
        declaredVersion:
          type: string
          nullable: true
          maxLength: 128
          description: >-
            The plugin's own version, from its manifest's "version". Optional and additive - sent on every session
            so a self-updating plugin reports its new version without re-enrolling.
        sdk:
          $ref: "#/components/schemas/PluginSdkUsage"
    PluginSdkUsage:
      type: object
      description: >-
        Mirrors MacroDeck.Plugin.Protocol.Handshake.PluginSdkUsage. What the plugin was built against, and which
        deprecated APIs it actually uses. Optional and additive - a plugin that omits it still negotiates normally
        and is reported as compatibility source "unknown".
      required: [sdkVersion]
      properties:
        sdkVersion:
          type: string
          maxLength: 128
          description: The version of MacroDeck.Sdk the plugin was compiled against.
        deprecatedApis:
          type: array
          nullable: true
          maxItems: 64
          items:
            type: string
            maxLength: 256
          description: >-
            Documentation comment ids of the deprecated APIs the plugin references, captured at build time by the
            SDK's usage-manifest generator. Null and empty mean different things and must never be conflated - null
            is "not reported", so the host can only infer from sdkVersion; empty is "reported, and none are used",
            which the host treats as confirmed.
        truncated:
          type: boolean
          description: True when the generator hit its cap, so the list is a floor rather than the complete set.
    CompatibilityFinding:
      type: object
      description: Mirrors MacroDeck.Plugin.Protocol.Compatibility.CompatibilityFinding.
      required: [diagnosticId, source, severity, subject, guidance]
      properties:
        diagnosticId:
          type: string
          description: The analyzer id this corresponds to, e.g. MDP5002.
        source:
          type: string
          enum: [confirmed, negotiated, inferred, unknown]
          description: The evidence behind the finding. An "inferred" finding says the plugin may be affected.
        severity:
          type: string
          enum: [info, warning, error]
        subject:
          type: string
          description: A deprecated API's display name, a capability kind, or "protocol".
        guidance:
          type: string
        deprecatedIn:
          type: string
          nullable: true
        removedIn:
          type: string
          nullable: true
        replacement:
          type: string
          nullable: true
        migrationUrl:
          type: string
          nullable: true
    PluginCompatibilityReport:
      type: object
      description: Mirrors MacroDeck.Plugin.Protocol.Compatibility.PluginCompatibilityReport.
      required: [state, usageSource, negotiatedProtocolVersion, findings]
      properties:
        state:
          type: string
          enum:
            [
              compatible,
              deprecated_apis,
              update_recommended,
              update_required,
              partially_incompatible,
              incompatible,
            ]
        usageSource:
          type: string
          enum: [confirmed, negotiated, inferred, unknown]
        negotiatedProtocolVersion:
          type: integer
        sdkVersion:
          type: string
          nullable: true
        usageTruncated:
          type: boolean
        findings:
          type: array
          maxItems: 128
          items:
            $ref: "#/components/schemas/CompatibilityFinding"
    PluginSessionResponse:
      type: object
      description: Mirrors MacroDeck.Plugin.Protocol.Handshake.PluginSessionResponse.
      required:
        - sessionId
        - sessionToken
        - negotiatedVersion
        - capabilities
        - limits
        - timeouts
      properties:
        sessionId:
          type: string
        sessionToken:
          type: string
        negotiatedVersion:
          type: integer
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/CapabilityNegotiationResult"
        limits:
          $ref: "#/components/schemas/PluginProtocolLimitsDescriptor"
        timeouts:
          $ref: "#/components/schemas/PluginProtocolTimeoutsDescriptor"
        compatibility:
          $ref: "#/components/schemas/PluginCompatibilityReport"